Update requirements.txt
Most notable
dataclasses==0.8 REMOVED
dataclasses==0.8 is not available for Python 3.10.16. This makes sense because the dataclasses module was introduced in Python 3.7, and for Python 3.10, it is part of the standard library. That means you don't need to install it separately.
numpy<2.0
networkx was using np.float_, which was removed in NumPy 2.0.
scipy>=1.15.2
scipy==1.5.4 is trying to install but fails due to dependency issues, mainly with numpy==1.17.3, which is outdated and incompatible with Python 3.10.16.